🎨 style : stack CSS 수정
